+++ Fetch/Parsable.swift
@@ -22,21 +22,3 @@ public protocol Parsable {
     static func parse(response: Response, errorParser: ErrorParsing.Type?) -> Result<Self>
 
 }
-
-public struct Response {
-
-    /// The data received
-    public let data: Data?
-
-    /// The HTTP status code received
-    public let status: Int
-
-    /// The HTTP headers received
-    public let headers: [String: String]?
-
-    /// A userInfo dictionary attached to the request.
-    public let userInfo: [String: Any]?
-
-    /// The request this response was created from.
-    public let originalRequest: Request
-}
